FHIR_COMBINED/FHIR_dashboard/backend/services/ml-hotspots/requirements.txt
pandas==2.1.4
numpy==1.26.2
mysql-connector-python==8.2.0
python-dotenv==1.0.0
scikit-learn==1.3.2

